unify profile progress trackers
