+++ app/src/main/res/values-ar/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">لم يواجه بعد</string>
     <string name="bestiary_qty_header">كمية</string>
     <string name="bestiary_chance_header">احتمالية</string>
-    <string name="bestiary_encountered_progress">%1$d / %2$d تمت مواجهته</string>
+    <string name="label_completion_progress">%1$d / %2$d %3$s</string>
+    <string name="achievements_progress_bar">obtained</string>
+    <string name="bestiary_progress_bar">تمت مواجهته</string>
     <string name="bestiary_always_drops">دوماً يعطي</string>
     <string name="bestiary_rare_drops">نادراً يعطي</string>
     <string name="bestiary_common_loot">غنائم شائعة</string>
