+++ app/src/main/res/values-de/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">Noch nicht angetroffen</string>
     <string name="bestiary_qty_header">Menge</string>
     <string name="bestiary_chance_header">Chance</string>
-    <string name="bestiary_encountered_progress">%1$d / %2$d angetroffen</string>
+    <string name="label_completion_progress">%1$d / %2$d %3$s</string>
+    <string name="achievements_progress_bar">erhalten</string>
+    <string name="bestiary_progress_bar">angetroffen</string>
     <string name="bestiary_always_drops">Lässt immer fallen</string>
     <string name="bestiary_rare_drops">Seltene Beute</string>
     <string name="bestiary_common_loot">Gewöhnliche Beute</string>
