+++ app/src/main/res/values-es/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">Aún no encontrado</string>
     <string name="bestiary_qty_header">Qty</string> <!-- untranslated -->
     <string name="bestiary_chance_header">Chance</string> <!-- untranslated -->
-    <string name="bestiary_encountered_progress">%1$d / %2$d encountered</string> <!-- untranslated -->
+    <string name="label_completion_progress">%1$d / %2$d %3$s</string>
+    <string name="achievements_progress_bar">obtenidos</string>
+    <string name="bestiary_progress_bar">encountered</string> <!-- untranslated -->
     <string name="bestiary_always_drops">Siempre cae</string>
     <string name="bestiary_rare_drops">Objetos raros</string>
     <string name="bestiary_common_loot">Botín común</string>
