+++ app/src/main/res/values-fr/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">Pas encore rencontré</string>
     <string name="bestiary_qty_header">Qté</string>
     <string name="bestiary_chance_header">Chance</string>
-    <string name="bestiary_encountered_progress">%1$d / %2$d croisés</string>
+    <string name="label_completion_progress">%1$d / %2$d %3$s</string>
+    <string name="achievements_progress_bar">obtenus</string>
+    <string name="bestiary_progress_bar">croisés</string>
     <string name="bestiary_always_drops">Donne Toujours</string>
     <string name="bestiary_rare_drops">Donne Rarement</string>
     <string name="bestiary_common_loot">Butin Commun</string>
