+++ app/src/main/res/values-it/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">Non ancora incontrato</string>
     <string name="bestiary_qty_header">Q.tà</string>
     <string name="bestiary_chance_header">Possibilità</string>
-    <string name="bestiary_encountered_progress">%1$d / %2$d incontrati</string>
+    <string name="label_completion_progress">%1$d / %2$d %3$s</string>
+    <string name="achievements_progress_bar">ottenuti</string>
+    <string name="bestiary_progress_bar">incontrati</string>
     <string name="bestiary_always_drops">Bottino garantito</string>
     <string name="bestiary_rare_drops">Bottino raro</string>
     <string name="bestiary_common_loot">Bottino tipico</string>
