+++ app/src/main/res/values-ru/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">Ещё не встречен</string>
     <string name="bestiary_qty_header">Количество</string>
     <string name="bestiary_chance_header">Шанс</string>
-    <string name="bestiary_encountered_progress">%1$d / %2$d открыто</string>
+    <string name="label_completion_progress">%1$d / %2$d %3$s</string>
+    <string name="achievements_progress_bar">получено</string>
+    <string name="bestiary_progress_bar">открыто</string>
     <string name="bestiary_always_drops">Всегда выпадает</string>
     <string name="bestiary_rare_drops">Редкая добыча</string>
     <string name="bestiary_common_loot">Обычная добыча</string>
