+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">尚未遭遇</string>
     <string name="bestiary_qty_header">数量</string>
     <string name="bestiary_chance_header">几率</string>
-    <string name="bestiary_encountered_progress">已遇见 %1$d / %2$d</string>
+    <string name="label_completion_progress">%3$s %1$d / %2$d</string>
+    <string name="achievements_progress_bar">已获取</string>
+    <string name="bestiary_progress_bar">"已遇见 "</string>
     <string name="bestiary_always_drops">必定掉落</string>
     <string name="bestiary_rare_drops">稀有掉落</string>
     <string name="bestiary_common_loot">常见战利品</string>
