+++ app/src/main/res/values/strings.xml
@@ -67,7 +67,9 @@
     <string name="bestiary_not_encountered">Not yet encountered</string>
     <string name="bestiary_qty_header">Qty</string>
     <string name="bestiary_chance_header">Chance</string>
-    <string name="bestiary_encountered_progress">%1$d / %2$d encountered</string>
+    <string name="label_completion_progress">%1$d / %2$d %3$s</string>
+    <string name="achievements_progress_bar">obtained</string>
+    <string name="bestiary_progress_bar">encountered</string>
     <string name="bestiary_always_drops">Always Drops</string>
     <string name="bestiary_rare_drops">Rare Drops</string>
     <string name="bestiary_common_loot">Common Loot</string>
